+++ packages/core/src/configuration/index.ts
@@ -1,0 +1,3 @@
+export { AeroGearConfiguration } from "./AerogearConfiguration";
+export { ConfigurationParser } from "./ConfigurationParser";
+export { ServiceConfiguration } from "./ServiceConfiguration";
