+++ packages/core/src/metrics/model/AppMetrics.ts
@@ -1,9 +1,7 @@
-interface AppMetrics {
+export interface AppMetrics {
 
   appId: string;
   sdkVersion: string;
   appVersion: string;
 
 }
-
-export default AppMetrics;
