+++ packages/core/src/metrics/model/DeviceMetrics.ts
@@ -1,8 +1,6 @@
-interface DeviceMetrics {
+export interface DeviceMetrics {
 
   platform: string;
   platformVersion: string;
 
 }
-
-export default DeviceMetrics;
