+++ packages/core/src/metrics/publisher/index.ts
@@ -1,0 +1,2 @@
+export { MetricsPublisher } from "./MetricsPublisher";
+export { MetricsNetworkPublisher } from "./NetworkMetricsPublisher";
